refactor: DiffCallback을 별도의 클래스로 분리
