+++ app/src/main/java/com/june0122/wakplus/utils/DiffCallback.kt
@@ -1,0 +1,25 @@
+package com.june0122.wakplus.utils
+
+import androidx.recyclerview.widget.DiffUtil
+
+open class DiffCallback<T>(
+    val oldItems: List<T>,
+    val newItems: List<T>
+) : DiffUtil.Callback() {
+    override fun getOldListSize(): Int = oldItems.size
+    override fun getNewListSize(): Int = newItems.size
+
+    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+        val oldItem = oldItems[oldItemPosition]
+        val newItem = newItems[newItemPosition]
+
+        return oldItem == newItem
+    }
+
+    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+        val oldItem = oldItems[oldItemPosition]
+        val newItem = newItems[newItemPosition]
+
+        return oldItem == newItem
+    }
+}
